LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/mathml/presentation-markup/fractions/frac-linethickness-002-expected.txt









PASS Positive
PASS Negative
PASS Percentage
FAIL Named space assert_approx_equals: expected 100 +/- 2 but got 3.875
PASS Calc() expression
FAIL Dynamic linethickness assert_approx_equals: attach expected 400 +/- 2 but got 100
